+++ crates/settings/src/vscode_import.rs
@@ -584,6 +584,7 @@ impl VsCodeSettings {
     fn status_bar_settings_content(&self) -> Option<StatusBarSettingsContent> {
         skip_default(StatusBarSettingsContent {
             show: self.read_bool("workbench.statusBar.visible"),
+            active_file: None,
             active_language_button: None,
             cursor_position_button: None,
             line_endings_button: None,